+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This is the first release to use a changelog.
 ### Added
 
 - Add template class `strided_view<T>` for non-contiguous evenly-strided data
+- `errors.h`: Use `string_view` for `_error_message_impl`
 - Extend support for coal `CollisionGeometry` objects
   - Add support for `coal::GEOM_ELLIPSOID`
   - Add support for `coal::GEOM_CONVEX`